Mockups - Restrict access to categories routes and management links

+++ resources/views/mockups/index.blade.php
@@ -5,19 +5,25 @@
 	<div class="sortable panel panel-default list-group">
 		@forelse ($categories as $category)
 			<div class="list-group-item" data-id="{{ $category->id }}">
-				<i class=" glyphicon glyphicon-th"></i>
+				@if (Auth::check())
+					<i class=" glyphicon glyphicon-th"></i>
+				@endif
 				<a class="name" href="{{ action('MockupCategoryController@show', [$project->slug, $category->id]) }}">{{ $category->name }}</a>
-				{!! Form::open(['url' => action('MockupCategoryController@destroy', [$project->slug, $category->id]), 'method' => 'DELETE', 'class' => 'delete-element pull-right']) !!}
-					<div class="form-group">
-					    {!! Form::submit('×', ['class' => 'close'])!!}
-					</div>
-			    {!!  Form::close() !!}
+				@if (Auth::check())
+					{!! Form::open(['url' => action('MockupCategoryController@destroy', [$project->slug, $category->id]), 'method' => 'DELETE', 'class' => 'delete-element pull-right']) !!}
+						<div class="form-group">
+						    {!! Form::submit('×', ['class' => 'close'])!!}
+						</div>
+				    {!!  Form::close() !!}
+				@endif
 			</div>
 		@empty
 			{!! trans('mockup.no_categories') !!}
 		@endforelse
 	</div>
-	<a href="{{ action('MockupController@create', [$project->slug])}}" class="btn btn-primary">{!! trans('mockup.create_mockup') !!}</a>
+	@if (Auth::check())
+		<a href="{{ action('MockupController@create', [$project->slug])}}" class="btn btn-primary">{!! trans('mockup.create_mockup') !!}</a>
+	@endif
 @endsection
 @section('footer_js')
 	<script>
