+++ resources/views/mockups/list.blade.php
@@ -5,20 +5,26 @@
 	<div class="sortable panel panel-default list-group">
 		@forelse ($mockups as $mockup)
 			<div class="list-group-item {{ $mockup->color }}" data-id="{{ $mockup->id }}">
-				<i class=" glyphicon glyphicon-th"></i>
+				@if (Auth::check())
+					<i class=" glyphicon glyphicon-th"></i>
+				@endif
 				<span class="format-{{ $mockup->format }}"></span>
 				<a class="name" href="{{ action('MockupController@show', [$project->slug, $mockup->id]) }}">{{ $mockup->name }}</a>
-				{!! Form::open(['url' => action('MockupController@destroy', [$project->slug, $mockup->id]), 'method' => 'DELETE', 'class' => 'delete-element pull-right']) !!}
-					<div class="form-group">
-					    {!! Form::submit('×', ['class' => 'close'])!!}
-					</div>
-			    {!!  Form::close() !!}
+				@if (Auth::check())
+					{!! Form::open(['url' => action('MockupController@destroy', [$project->slug, $mockup->id]), 'method' => 'DELETE', 'class' => 'delete-element pull-right']) !!}
+						<div class="form-group">
+						    {!! Form::submit('×', ['class' => 'close'])!!}
+						</div>
+				    {!!  Form::close() !!}
+				@endif
 			</div>
 		@empty
 			{!! trans('mockup.no_mockups') !!}
 		@endforelse
 	</div>
-	<a href="{{ action('MockupController@create', [$project->slug])}}" class="btn btn-primary">{!! trans('mockup.create_mockup') !!}</a>
+	@if (Auth::check())
+		<a href="{{ action('MockupController@create', [$project->slug])}}" class="btn btn-primary">{!! trans('mockup.create_mockup') !!}</a>
+	@endif
 @endsection
 @section('footer_js')
 	<script>
